requirements.txt
pyswisseph>=2.10.3
numpy>=1.26.0
fastapi>=0.100.0
uvicorn>=0.24.0
anthropic>=0.20.0
pytest>=7.0.0
httpx>=0.25.0
pydantic>=2.0.0
psycopg2-binary>=2.9.0
scipy>=1.11.0
scikit-learn>=1.3.0
